app/pages: init note page
There is still quite a bit to deal with here but the idea is set.
What I hope to do is to open a note from the local database and then
update it when necessary.
The database would be checked at startup and whenever a note is modified
to sync it.